fix: add throttling behaviour to navigation
